fix main

+++ pytorch_dqn/main.py
@@ -58,8 +58,10 @@ if config.controller:
 eval_folder = os.path.abspath(os.path.dirname(__file__) + "/../" + config.evaluation_directory_name)
 
 # Cleaning the evaluation folder
-shutil.rmtree(eval_folder)
-os.mkdir(eval_folder)
+if os.path.exists(eval_folder) and os.path.isdir(eval_folder):
+    shutil.rmtree(eval_folder)
+    shutil.rmtree(eval_folder)
+    os.mkdir(eval_folder)
 
 # Initializing evaluation
 evaluator_frames = ev_frames("dqn")
